+++ IspDataDb/dbo/Tables/EmployeeLeave.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[EmployeeLeave] (
     [IdNo]               INT      IDENTITY (1, 1) NOT NULL,
-    [EmployeeIdNo]       SMALLINT NULL,
+    [EmployeeIdNo]       INT      NULL,
     [LeaveIdNo]          SMALLINT NOT NULL,
     [DateStart]          DATE     NULL,
     [DateEnd]            DATE     NULL,
@@ -8,6 +8,8 @@
     [HRApproved]         BIT      NULL,
     CONSTRAINT [PK_EmployeeLeave] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
