+++ IspDataDb/dbo/Tables/ProductCategory.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[ProductCategory] (
-    [IdNo]                   INT            IDENTITY (1, 1) NOT NULL,
+    [IdNo]                   SMALLINT       IDENTITY (1, 1) NOT NULL,
     [ProductCategoryCode]    VARCHAR (10)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ProductCategoryName]    VARCHAR (50)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ProductCategoryNameAra] NVARCHAR (100) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
@@ -7,4 +7,6 @@
     [DateTimeStamp]          ROWVERSION     NULL,
     CONSTRAINT [PK_Category] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
